requirements.txt
streamlit
psycopg2-binary
pandas
plotly
bcrypt
python-dotenv
openpyxl
